pyproject.toml
[project]
name = "001-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.5.1",
    "gradio-pdf>=0.0.24",
    "ipykernel>=7.1.0",
    "langchain>=1.2.8",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.7",
    "openai>=2.16.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
]
